Inserindo os models

+++ src/main/java/com/AnormaisEBruna/petshop/repositories/UserRepository.java
@@ -4,7 +4,7 @@ import com.AnormaisEBruna.petshop.models.UserModel;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.CrudRepository;
 
-public interface UserRepository extends CrudRepository<UserModel, Long> {
+public interface UserRepository extends CrudRepository<UserModel, Integer> {
     @Query("SELECT u FROM UserModel as u WHERE u.email LIKE :email AND :password LIKE :password")
     UserModel findByCredentials(String email, String password);
 }
